viacheslav (#13)
* [Refactor] improve code readibility

* [Feature] create profile-service module
[Feature] implement Feign client in authentication-service
[Feature] implement WireMock configuration for Feign tests
[Feature] implement MapStruct for mapping between entities and DTOs

* [Feature] add controller level validation in profile and auth services

+++ authentication-service/src/main/java/com/example/authentication/repository/TokenRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface TokenRepository extends JpaRepository<Token, Integer> {
+public interface TokenRepository extends JpaRepository<Token, Long> {
 
     Optional<Token> findByAccount_Id(Long id);
